properties_per_type: extract some functions to secondary libs files
as that one file is already quite hairy
