+++ app/modules/entities/lib/editor/properties_per_type.js
@@ -104,46 +104,3 @@ for (const [ type, propertiesData ] of Object.entries(propertiesPerType)) {
     propertiesPerTypeAndCategory[type][category][property] = propertyData
   }
 }
-
-export const getSubentitiesTypes = property => {
-  const subentitiesTypes = []
-  Object.keys(infoboxPropertiesByType).forEach(type => {
-    const typeProps = infoboxPropertiesByType[type]
-    if (typeProps.includes(property) && type !== 'article') {
-      subentitiesTypes.push(pluralize(type))
-    }
-  })
-  return subentitiesTypes
-}
-
-export function getTypePropertiesPerCategory (entity) {
-  const { type } = entity
-  if (type === 'work' || type === 'serie') {
-    const entityPropertiesPerTypeAndCategory = deepClone(propertiesPerTypeAndCategory[type])
-    entityPropertiesPerTypeAndCategory.general = customizeAuthorProperties(entity, entityPropertiesPerTypeAndCategory.general)
-    return entityPropertiesPerTypeAndCategory
-  } else {
-    return propertiesPerTypeAndCategory[type]
-  }
-}
-
-function customizeAuthorProperties (entity, generalProperties) {
-  return flatMapKeyValues(generalProperties, ([ property, propertySettings ]) => {
-    if (property === 'wdt:P50') {
-      const customAuthorProperties = getWorkPreferredAuthorRolesProperties(entity)
-      let entries = customAuthorProperties.map(prop => [ prop, properties[prop] ])
-      if (!customAuthorProperties.includes('wdt:P50') && isNonEmptyArray(entity.claims['wdt:P50'])) {
-        entries = [
-          [ 'wdt:P50', propertySettings ],
-          ...entries
-        ]
-      }
-      return entries
-    } else if (propertySettings.contextual) {
-      // Contextual author properties might have possibly been introduced by getWorkPreferredAuthorRolesProperties
-      return []
-    } else {
-      return [ [ property, propertySettings ] ]
-    }
-  })
-}
